move timer definitions into TimerTypes
