+++ include/TimerTypes.h
@@ -5,6 +5,7 @@
 
 namespace aperi {
 
+// Add FunctionCreationProcessorTimerType from FunctionEvaluationProcessor.h
 enum class FunctionCreationProcessorTimerType {
     Instantiate,
     ComputeFunctionValues,
@@ -16,8 +17,28 @@ inline const std::map<FunctionCreationProcessorTimerType, std::string> function_
     {FunctionCreationProcessorTimerType::ComputeFunctionValues, "ComputeFunctionValues"},
     {FunctionCreationProcessorTimerType::NONE, "NONE"}};
 
-//
-// Add other timer types and maps here as needed
-//
+// Add NeighborSearchProcessorTimerType from NeighborSearchProcessor.h
+enum class NeighborSearchProcessorTimerType {
+    Instantiate,
+    KokkosDeepCopy,
+    CoarseSearch,
+    UnpackSearchResultsIntoField,
+    GhostNodeNeighbors,
+    CreateNodeSpheres,
+    CreateNodePoints,
+    ComputeKernelRadius,
+    NONE
+};
+
+inline const std::map<NeighborSearchProcessorTimerType, std::string> neighbor_search_processor_timer_names_map = {
+    {NeighborSearchProcessorTimerType::Instantiate, "Instantiate"},
+    {NeighborSearchProcessorTimerType::KokkosDeepCopy, "KokkosDeepCopy"},
+    {NeighborSearchProcessorTimerType::CoarseSearch, "CoarseSearch"},
+    {NeighborSearchProcessorTimerType::UnpackSearchResultsIntoField, "UnpackSearchResultsIntoField"},
+    {NeighborSearchProcessorTimerType::GhostNodeNeighbors, "GhostNodeNeighbors"},
+    {NeighborSearchProcessorTimerType::CreateNodeSpheres, "CreateNodeSpheres"},
+    {NeighborSearchProcessorTimerType::CreateNodePoints, "CreateNodePoints"},
+    {NeighborSearchProcessorTimerType::ComputeKernelRadius, "ComputeKernelRadius"},
+    {NeighborSearchProcessorTimerType::NONE, "NONE"}};
 
 }  // namespace aperi
